+++ Banjo/Source/Utility/Math/Vector2.h
@@ -3,9 +3,9 @@
 class Vector2
 {
 public:
-    float x, y;
+    int x, y;
     Vector2();
-    Vector2(float x, float y);
+    Vector2(int x, int y);
 
     Vector2 operator-=(Vector2 const& vector);
     Vector2 operator-(Vector2 const& vector) const;
@@ -13,8 +13,8 @@ public:
     Vector2 operator+(Vector2 const& vector) const;
     Vector2 operator*=(Vector2 const& vector);
     Vector2 operator*(Vector2 const& vector) const;
-    Vector2 operator*=(float const& scalar);
-    Vector2 operator*(float const& scalar) const;
+    Vector2 operator*=(int const& scalar);
+    Vector2 operator*(int const& scalar) const;
 
     static float dot_product(Vector2 const& vectorA, Vector2 const& vectorB);
