+++ Banjo/Source/Utility/Math/Vector3.h
@@ -5,7 +5,7 @@ class Vector3
 {
 public:
     Vector3();
-    Vector3(float x,  float y,  float z);
+    Vector3(int x,  int y,  int z);
 
     Vector3 operator+(Vector3 const& vector) const;
     Vector3 operator +=(Vector3 const& vector);
@@ -16,11 +16,11 @@ public:
     Vector3 operator*(Vector3 const& vector) const;
     Vector3 operator *=(Vector3 const& vector);
 
-    Vector3 operator*(float scalar) const;
-    Vector3 operator*=(float scalar);
+    Vector3 operator*(int scalar) const;
+    Vector3 operator*=(int scalar);
 
     static float dot_product(Vector3 const& vectorA, Vector3 const& vectorB);
     static Vector3 cross_product(Vector3 const& vectorA, Vector3 const& vectorB);
 
-    float x, y, z;
+    int x, y, z;
 };
